pyproject.toml
[tool.poetry]
name = "phone-variable-scroll"
version = "0.1.0"
description = ""
authors = ["John Ciubuc <john.ciubuc@ttuhsc.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
pure-python-adb = "^0.3.0-alpha.0"
python-uinput = "^0.11.2"
spyder = "^5.3.0"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]
python-uinput = "^0.11.2"
pure-python-adb = "^0.3.0-alpha.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
